requirements/common.txt
# Django
Django>=1.11, <2.0
django-extensions>=1.7.4

# Email
django-disposable-email-checker~=1.2

# Rest Framework
djangorestframework>=3.6.2
djangorestframework-camel-case>=0.2.0
django-rest-swagger>=2.1.2


# Deployment
uwsgi
